+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+- [PR-72](https://github.com/itk-dev/ai-screening/pull/72)
+  Added cache context
 - [PR-71](https://github.com/itk-dev/ai-screening/pull/71)
   - Added link to chart
   - Changed display of form submission values